pyproject.toml
[project]
name = "bnn"
version = "0.1.0"
description = ""
authors = [
    {name = "anon"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytz (>=2025.2,<2026.0)",
    "cupy-cuda12x (>=13.4.1,<14.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "ucimlrepo (>=0.0.7,<0.0.8)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "lightning (>=2.5.1.post0,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
